@@ -1,193 +1,20 @@
-process.env.NODE_ENV = 'development'
+#!/usr/bin/env node
 
-const detect = require('detect-port')
-const clearConsole = require('react-dev-utils/clearConsole')
-const getProcessForPort = require('react-dev-utils/getProcessForPort')
-const formatWebpackMessages = require('react-dev-utils/formatWebpackMessages')
-const openBrowser = require('react-dev-utils/openBrowser')
-const prompt = require('react-dev-utils/prompt')
-const webpack = require('webpack')
-const historyApiFallback = require('connect-history-api-fallback')
-// const httpProxyMiddleware = require('http-proxy-middleware')
-const WebpackDevServer = require('webpack-dev-server')
-const chalk = require('chalk')
-const paths = require('../config/paths')
-const getConfig = require('../utils/getConfig')
-const applyWebpackConfig = require('../utils/applyWebpackConfig')
+import { fork } from 'child_process'
 
-const DEFAULT_PORT = process.env.PORT ? parseInt(process.env.PORT, 10) : 8000
-const isInteractive = process.stdout.isTTY
-let compiler
-
-const argv = require('yargs')
-  .usage('Usage: bee server [options]')
-  .option('open', {
-    type: 'boolean',
-    describe: 'Open url in browser after started',
-    default: true
-  })
-  .help('h')
-  .argv
-
-let rcConfig
-try {
-  rcConfig = getConfig(process.env.NODE_ENV)
-} catch (e) {
-  console.log(chalk.red('Failed to parse .beerc config.'))
-  console.log()
-  console.log(e.message)
-  process.exit(1)
-}
-
-const config = applyWebpackConfig(
-  require('../config/webpack.dev.config'),
-  process.env.NODE_ENV
-)
-
-function setupCompiler (host, port, protocol) {
-  compiler = webpack(config)
-
-  compiler.plugin('invalid', function () {
-    if (isInteractive) {
-      clearConsole()
-    }
-    console.log('Compiling...')
-  })
-
-  let isFirstCompile = true
-  compiler.plugin('done', function (stats) {
-    if (isInteractive) {
-      clearConsole()
-    }
-
-    const messages = formatWebpackMessages(stats.toJson({}, true))
-    const isSuccessful = !messages.errors.length && !messages.warnings.length
-    const showInstructions = isSuccessful && (isInteractive || isFirstCompile)
-
-    applyWebpackConfig.warnIfExists()
-
-    if (isSuccessful) {
-      console.log(chalk.green('Compiled successfully!'))
-    }
-
-    if (showInstructions) {
-      console.log()
-      console.log('The app is running at:')
-      console.log()
-      console.log('  ' + chalk.cyan(protocol + '://' + host + ':' + port + '/'))
-      console.log()
-      console.log('Note that the development build is not optimized.')
-      console.log(`To create a production build, use ${chalk.cyan('npm run build')}.`)
-      console.log()
-      isFirstCompile = false
-    }
-
-    // If errors exist, only show errors.
-    if (messages.errors.length) {
-      console.log(chalk.red('Failed to compile.'))
-      console.log()
-      messages.errors.forEach(message => {
-        console.log(message)
-        console.log()
-      })
-      return
-    }
-
-    // Show warnings if no errors were found.
-    if (messages.warnings.length) {
-      console.log(chalk.yellow('Compiled with warnings.'))
-      console.log()
-      messages.warnings.forEach(message => {
-        console.log(message)
-        console.log()
-      })
-      // Teach some ESLint tricks.
-      console.log('You may use special comments to disable some warnings.')
-      console.log('Use ' + chalk.yellow('// eslint-disable-next-line') + ' to ignore the next line.')
-      console.log('Use ' + chalk.yellow('/* eslint-disable */') + ' to ignore all warnings in a file.')
+function start () {
+  const p = fork(`${__dirname}/server`, process.argv.slice(2))
+  p.on('message', (data) => {
+    if (data === 'RESTART') {
+      p.kill('SIGINT')
+      start()
     }
   })
 }
 
-function addMiddleware (devServer) {
-  const proxy = require(paths.appPackageJson).proxy
-  devServer.use(historyApiFallback({
-    disableDotRule: true,
-    htmlAcceptHeaders: proxy ? ['text/html'] : ['text/html', '*/*']
-  }))
-  // TODO: proxy index.html, ...
-  devServer.use(devServer.middleware)
+if (!process.send) {
+  start()
+} else {
+  require('./createServer')
 }
 
-function runDevServer (host, port, protocol) {
-  // Run DEV server for hot-reloading
-  const devServer = new WebpackDevServer(compiler, {
-    compress: true,
-    clientLogLevel: 'none',
-    contentBase: paths.appPublic,
-    headers: {
-      'Access-Control-Allow-Origin': '*',
-      'Access-Control-Expose-Headers': 'SourceMap,X-SourceMap'
-    },
-    hot: true,
-    publicPath: config.output.publicPath,
-    quiet: true,
-    watchOptions: {
-      aggregateTimeout: 300,
-      poll: false,
-      ignored: /node_modules/
-    },
-    https: protocol === 'https',
-    host,
-    proxy: rcConfig.proxy
-  })
-
-  addMiddleware(devServer)
-
-  devServer.listen(port, (err) => {
-    if (err) {
-      return console.log(err)
-    }
-
-    if (isInteractive) {
-      clearConsole()
-    }
-    console.log(chalk.cyan('Starting the development server...'))
-    console.log()
-
-    if (argv.open) {
-      openBrowser(protocol + '://' + host + ':' + port + '/')
-    }
-  })
-}
-
-function run (port) {
-  const protocol = process.env.HTTPS === 'true' ? 'https' : 'http'
-  const host = process.env.HOST || 'localhost'
-  setupCompiler(host, port, protocol)
-  runDevServer(host, port, protocol)
-}
-
-detect(DEFAULT_PORT).then((port) => {
-  if (port === DEFAULT_PORT) {
-    run(port)
-    return
-  }
-
-  if (isInteractive) {
-    clearConsole()
-    const existingProcess = getProcessForPort(DEFAULT_PORT)
-    const question =
-      chalk.yellow('Something is already running on port ' + DEFAULT_PORT + '.' +
-        ((existingProcess) ? ' Probably:\n  ' + existingProcess : '')) +
-      '\n\nWould you like to run the app on another port instead?'
-
-    prompt(question, true).then(shouldChangePort => {
-      if (shouldChangePort) {
-        run(port)
-      }
-    })
-  } else {
-    console.log(chalk.red(`Something is already running on port ${DEFAULT_PORT}.`))
-  }
-})
