@@ -1,24 +1,26 @@
-const path = require('path')
-const fs = require('fs')
-
-const appDirectory = fs.realpathSync(process.cwd())
-
-function resolveApp (relativePath) {
-  return path.resolve(appDirectory, relativePath)
-}
+import { resolve } from 'path'
+import { realpathSync } from 'fs'
 
 function resolveOwn (relativePath) {
-  return path.resolve(__dirname, relativePath)
+  return resolve(__dirname, relativePath)
 }
 
-module.exports = {
-  appSrc: resolveApp('src'),
-  appBuild: resolveApp('dist'),
-  appPublic: resolveApp('public'),
-  appPackageJson: resolveApp('package.json'),
-  appNodeModules: resolveApp('node_modules'),
-  ownNodeModules: resolveOwn('../node_modules'),
-  appDirectory,
-  resolveApp,
-  resolveOwn
+export default function getPaths (cwd) {
+  const appDirectory = realpathSync(cwd)
+
+  function resolveApp (relativePath) {
+    return resolve(appDirectory, relativePath)
+  }
+
+  return {
+    appSrc: resolveApp('src'),
+    appBuild: resolveApp('dist'),
+    appPublic: resolveApp('public'),
+    appPackageJson: resolveApp('package.json'),
+    appNodeModules: resolveApp('node_modules'),
+    ownNodeModules: resolveOwn('../node_modules'),
+    appDirectory,
+    resolveApp,
+    resolveOwn
+  }
 }
