@@ -1,6 +1,6 @@
-const ExtractTextPlugin = require('extract-text-webpack-plugin')
+import ExtractTextPlugin from 'extract-text-webpack-plugin'
 
-exports.cssLoaders = function (options) {
+function cssLoaders (options) {
   options = options || {}
   // generate loader string to be used with extract text plugin
   function generateLoaders (loaders) {
@@ -37,7 +37,7 @@ exports.cssLoaders = function (options) {
 }
 
 // Generate loaders for standalone style files (outside of .vue)
-exports.styleLoaders = function (options) {
+function styleLoaders (options) {
   let output = []
   const loaders = exports.cssLoaders(options)
   for (const extension in loaders) {
@@ -48,4 +48,9 @@ exports.styleLoaders = function (options) {
     })
   }
   return output
+}
+
+export default {
+  cssLoaders,
+  styleLoaders
 }
