@@ -1,8 +1,9 @@
-const fs = require('fs')
-const stripJsonComments = require('strip-json-comments')
-const isPlainObject = require('is-plain-object')
-const parseJSON = require('parse-json-pretty')
-const paths = require('../config/paths')
+import { existsSync, readFileSync } from 'fs'
+import stripJsonComments from 'strip-json-comments'
+import isPlainObject from 'is-plain-object'
+import parseJSON from 'parse-json-pretty'
+import getPaths from '../config/paths'
+
 require('./registerBabel')
 
 function merge (oldObj, newObj) {
@@ -17,30 +18,46 @@ function merge (oldObj, newObj) {
   }
 }
 
-function getConfig (configFile) {
+function getConfig (configFile, paths) {
   const rcConfig = paths.resolveApp(configFile)
   const jsConfig = paths.resolveApp(`${configFile}.js`)
 
-  if (fs.existsSync(rcConfig)) {
-    return parseJSON(stripJsonComments(fs.readFileSync(rcConfig, 'utf-8')), `./${configFile}`)
-  } else if (fs.existsSync(jsConfig)) {
-    return require(jsConfig)
+  if (existsSync(rcConfig)) {
+    // return parseJSON(stripJsonComments(readFileSync(rcConfig, 'utf-8')), './beerc')
+    return parseJSON(stripJsonComments(readFileSync(rcConfig, 'utf-8')), `./${configFile}`)
+  } else if (existsSync(jsConfig)) {
+    return require(jsConfig);  // eslint-disable-line
   } else {
     return {}
   }
 }
 
-function realGetConfig (configFile, env = 'development') {
-  const config = getConfig(configFile)
+function replaceNpmVariables (value, pkg) {
+  if (typeof value === 'string') {
+    return value
+      .replace('$npm_package_name', pkg.name)
+      .replace('$npm_package_version', pkg.version)
+  } else {
+    return value
+  }
+}
+
+export function realGetConfig (configFile, env, pkg = {}, paths) {
+  env = env || 'development'
+  const config = getConfig(configFile, paths)
   if (config.env) {
     if (config.env[env]) merge(config, config.env[env])
     delete config.env
   }
-  return config
+
+  return Object.keys(config).reduce((memo, key) => {
+    memo[key] = replaceNpmVariables(config[key], pkg)
+    return memo
+  }, {})
 }
 
-module.exports = function () {
-  return realGetConfig('.beerc', process.env.NODE_ENV)
+export default function (env, cwd) {
+  const paths = getPaths(cwd)
+  const pkg = JSON.parse(readFileSync(paths.appPackageJson, 'utf-8'))
+  return realGetConfig('.roadhogrc', env, pkg, paths)
 }
-
-module.exports.realGetConfig = realGetConfig
