@@ -1,13 +1,11 @@
-const path = require('path')
-const assert = require('assert')
-const glob = require('glob')
-const paths = require('../config/paths')
-const getConfig = require('./getConfig')
+import { basename, sep } from 'path'
+import assert from 'assert'
+import glob from 'glob'
 
 const DEFAULT_ENTRY = './src/index.js'
 
 function getEntry (filePath, isProduction) {
-  const key = path.basename(filePath, '.js')
+  const key = basename(filePath, '.js')
   const value = isProduction
     ? [filePath]
     : [
@@ -19,7 +17,7 @@ function getEntry (filePath, isProduction) {
   }
 }
 
-function getFiles (entry, cwd) {
+export function getFiles (entry, cwd) {
   if (Array.isArray(entry)) {
     return entry.reduce((memo, entryItem) => {
       return memo.concat(getFiles(entryItem, cwd))
@@ -27,29 +25,26 @@ function getFiles (entry, cwd) {
   } else {
     assert(
       typeof entry === 'string',
-      `getEntry/getFiles: entry type should be string, but got ${typeof entry}`
+      `getEntry/getFiles: entry type should be string, but got ${typeof entry}`,
     )
     const files = glob.sync(entry, {
       cwd
     })
     return files.map((file) => {
-      return (file.charAt(0) === '.') ? file : `.${path.sep}${file}`
+      return (file.charAt(0) === '.') ? file : `.${sep}${file}`
     })
   }
 }
 
-function getEntries (files, isProduction) {
+export function getEntries (files, isProduction) {
   return files.reduce((memo, file) => {
     return Object.assign(memo, getEntry(file, isProduction))
   }, {})
 }
 
-module.exports = function () {
-  const entry = getConfig().entry
+export default function (config, appDirectory) {
+  const entry = config.entry
   const isProduction = process.env.NODE_ENV === 'production'
-  const files = entry ? getFiles(entry, paths.appDirectory) : [DEFAULT_ENTRY]
+  const files = entry ? getFiles(entry, appDirectory) : [DEFAULT_ENTRY]
   return getEntries(files, isProduction)
 }
-
-module.exports.getFiles = getFiles
-module.exports.getEntries = getEntries
