@@ -1,4 +1,4 @@
-module.exports = function (define) {
+export default function normalizeDefine (define) {
   return Object.keys(define).reduce((memo, key) => {
     memo[key] = JSON.stringify(define[key])
     return memo
